requirements.txt
DateTime==4.4
PyJWT==2.3.0
uuid==1.30
gunicorn==20.1.0
Flask==2.1.2
boto==2.49.0